Finish reorganizing paint
